Output streaming integration tests (#342)
* Rename unary endpoint test file


* Update mocktail to 0.2.4-alpha


* test case: output_detector_no_detections()


* test case: output_detector_detections()


* test case: output_detector_client_error()


* refactor: replace assert! with assert_eq!


* Update output tests to use two detectors


* Rename streaming endpoint test file


* Change detector error message


* Add expected_messages to output_detectors_detections()


---------
