+++ Cargo.toml
@@ -90,7 +90,7 @@ tonic-build = "0.12.3"
 [dev-dependencies]
 axum-test = "17.1.0"
 faux = "0.1.12"
-mocktail = { version = "0.2.3-alpha" }
+mocktail = { version = "0.2.4-alpha" }
 rand = "0.9.0"
 test-log = "0.2.17"
