+++ tests/common/detectors.rs
@@ -18,6 +18,7 @@
 // Detector names
 pub const DETECTOR_NAME_ANGLE_BRACKETS_WHOLE_DOC: &str = "angle_brackets_detector_whole_doc";
 pub const DETECTOR_NAME_ANGLE_BRACKETS_SENTENCE: &str = "angle_brackets_detector_sentence";
+pub const DETECTOR_NAME_PARENTHESIS_SENTENCE: &str = "parenthesis_detector_sentence";
 pub const ANSWER_RELEVANCE_DETECTOR: &str = "answer_relevance_detector";
 pub const FACT_CHECKING_DETECTOR: &str = "fact_checking_detector";
 pub const PII_DETECTOR: &str = "pii_detector";
